requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
google-api-core==1.26.1
google-auth==1.27.1
google-cloud-pubsub==2.4.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.36.1
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.0
idna==2.10
libcst==0.3.17
mypy-extensions==0.4.3
packaging==20.9
priority==1.3.0
proto-plus==1.14.2
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
uvicorn==0.13.4
wsproto==1.0.0
